+++ src/components/DraftingInterface.tsx
@@ -1,17 +1,17 @@
 import { useState, useEffect, useCallback, useRef } from "react";
 import { useQuery } from "convex/react";
-import { api } from "../../../convex/_generated/api";
-import { CharacterPool } from "../CharacterPool/CharacterPool";
-import { TeamArea } from "../TeamArea/TeamArea";
-import { DraftControls } from "../DraftControls";
-import { CostTables } from "../CostTables";
-import { DraftProgress } from "../DraftProgress";
-import { TeamTest } from "../TeamTest/TeamTest";
-import { Contact } from "../Contact";
-import { DraftTimer } from "../DraftTimer";
-import { DraftingSettings } from "../DraftingSettings";
-import { Id } from "../../../convex/_generated/dataModel";
-import "./DraftingInterface.css";
+import { api } from "../../convex/_generated/api";
+import { CharacterPool } from "./CharacterPool";
+import { TeamArea } from "./TeamArea";
+import { DraftControls } from "./DraftControls";
+import { CostTables } from "./CostTables";
+import { DraftProgress } from "./DraftProgress";
+import { TeamTest } from "./TeamTest";
+import { Contact } from "./Contact";
+import { DraftTimer } from "./DraftTimer";
+import { DraftingSettings } from "./DraftingSettings";
+import { Id } from "../../convex/_generated/dataModel";
+import "../css/DraftingInterface.css";
 
 export type RuleSet = "memoryofchaos" | "apocalypticshadow";
 export type DraftMode = "4ban" | "6ban";
